+++ History.md
@@ -2,6 +2,7 @@
 - New "Implicit NotNull element cannot override CanBeNull in base type, nullability should be explicit" warning for return values and out parameters
 - New "Implicit NotNull result or out parameter overrides unknown nullability of external code" hint
 - Added suppression of "Base declaration has the same annotation" highlighting on implicitly nullable elements
+- Added exclusion of delegate BeginInvoke() method parameters and Invoke() / EndInvoke() results because their implicit nullability cannot be overridden with explicit annotations
 
 ### 2.2.0 ###
 - ReSharper 10.0 support